+++ internal/utils/privkey/privkey.go
@@ -2,25 +2,23 @@ package privkey
 
 import (
 	"errors"
-	"fmt"
 	"os"
 
 	"github.com/number571/go-peer/pkg/crypto/asymmetric"
 )
 
-func GetPrivKey(pKeyPath string) (asymmetric.IPrivKey, error) {
-	pubKeyPath := fmt.Sprintf("%s.pub", pKeyPath) //nolint:perfsprint
-	if _, err := os.Stat(pKeyPath); os.IsNotExist(err) {
+func GetPrivKey(pPrivKeyPath, pPubKeyPath string) (asymmetric.IPrivKey, error) {
+	if _, err := os.Stat(pPrivKeyPath); os.IsNotExist(err) {
 		privKey := asymmetric.NewPrivKey()
-		if err := os.WriteFile(pKeyPath, []byte(privKey.ToString()), 0600); err != nil {
+		if err := os.WriteFile(pPrivKeyPath, []byte(privKey.ToString()), 0600); err != nil {
 			return nil, errors.Join(ErrWritePrivateKey, err)
 		}
-		if err := os.WriteFile(pubKeyPath, []byte(privKey.GetPubKey().ToString()), 0600); err != nil {
+		if err := os.WriteFile(pPubKeyPath, []byte(privKey.GetPubKey().ToString()), 0600); err != nil {
 			return nil, errors.Join(ErrWritePublicKey, err)
 		}
 		return privKey, nil
 	}
-	privKeyStr, err := os.ReadFile(pKeyPath) //nolint:gosec
+	privKeyStr, err := os.ReadFile(pPrivKeyPath) //nolint:gosec
 	if err != nil {
 		return nil, errors.Join(ErrReadPrivateKey, err)
 	}
@@ -28,12 +26,12 @@ func GetPrivKey(pKeyPath string) (asymmetric.IPrivKey, error) {
 	if privKey == nil {
 		return nil, ErrInvalidPrivateKey
 	}
-	if _, err := os.Stat(pubKeyPath); os.IsNotExist(err) { //nolint:nestif
-		if err := os.WriteFile(pubKeyPath, []byte(privKey.GetPubKey().ToString()), 0600); err != nil {
+	if _, err := os.Stat(pPubKeyPath); os.IsNotExist(err) { //nolint:nestif
+		if err := os.WriteFile(pPubKeyPath, []byte(privKey.GetPubKey().ToString()), 0600); err != nil {
 			return nil, errors.Join(ErrWritePublicKey, err)
 		}
 	} else {
-		pubKeyStr, err := os.ReadFile(pubKeyPath) //nolint:gosec
+		pubKeyStr, err := os.ReadFile(pPubKeyPath) //nolint:gosec
 		if err != nil {
 			return nil, errors.Join(ErrReadPublicKey, err)
 		}