+++ pkg/api/adapters/http/client/errors.go
@@ -13,9 +13,8 @@ func (err *SError) Error() string {
 }
 
 var (
-	ErrBadRequest       = &SError{"bad request"}
-	ErrDecodeRequest    = &SError{"decode request"}
-	ErrDecodeResponse   = &SError{"decode response"}
-	ErrInvalidPublicKey = &SError{"invalid public key"}
-	ErrInvalidTitle     = &SError{"invalid title"}
+	ErrBadRequest     = &SError{"bad request"}
+	ErrDecodeRequest  = &SError{"decode request"}
+	ErrDecodeResponse = &SError{"decode response"}
+	ErrInvalidTitle   = &SError{"invalid title"}
 )