+++ pkg/api/kernel/client/proc/proc_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/number571/go-peer/pkg/crypto/scheme/layer2"
 )
 
-func TestXxx(t *testing.T) {
+func TestFriendsMapToList(t *testing.T) {
 	t.Parallel()
 
 	sortList := []string{"Alice", "Bob", "Carol"}